Restructuring login code from Sipity to global CAS
* Moves assertions for the sign in page to the global CAS login
support file 'spec/spec_support/cas_login.rb'
* Logout scenario now uses the same assertions for login page
